moves tool registration
The images tool now uses a structured content response. Following the suggestion in https://modelcontextprotocol.io/specification/2025-06-18/server/tools#output-schema, we also provide an Output Schema, enabling clients to validate the response. As a result of this work, the Images tool requires a slightly different registration approach. For this reason, registration can now be handled via a local "register" method in each tool. When all tools were registered externally, the type system is unable to infer the shape of each tools due to the overloaded union (resulting from only the Image tool having an OutputSchema export).